Fixed issue with command permission, fixed ChannelMonitor, removed unused code, fixed AutoInteraction not working.

+++ src/org/frostbite/karren/ChannelMonitor.java
@@ -10,6 +10,7 @@
 
 package org.frostbite.karren;
 
+import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.Category;
 import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.Invite;
@@ -27,15 +28,17 @@ public class ChannelMonitor extends Thread {
         kill = false;
         while(!kill){
             for(Guild guild : Karren.bot.getClient().getGuilds()){
-                List<Category> categoryList = guild.getCategoriesByName("temp voice channels", true);
-                if(categoryList.size()==1) {
-                    List<Invite> invites = guild.retrieveInvites().complete();
-                    for(VoiceChannel chan : categoryList.get(0).getVoiceChannels()){
-                        if(invites.stream().noneMatch(x-> Objects.requireNonNull(x.getChannel()).getId().equals(chan.getId())) || invites.isEmpty()){
-                            if(guild.getVoiceStates().stream().anyMatch(x-> Objects.equals(x.getChannel(), chan)))
-                                chan.createInvite().setMaxAge(3600).setMaxUses(1).setTemporary(false).setUnique(true).complete();
-                            else
-                                chan.delete().complete();
+                if(Objects.requireNonNull(guild.getMember(Karren.bot.client.getSelfUser())).hasPermission(Permission.MANAGE_SERVER)) {
+                    List<Category> categoryList = guild.getCategoriesByName("temp voice channels", true);
+                    if (categoryList.size() == 1) {
+                        List<Invite> invites = guild.retrieveInvites().complete();
+                        for (VoiceChannel chan : categoryList.get(0).getVoiceChannels()) {
+                            if (invites.stream().noneMatch(x -> Objects.requireNonNull(x.getChannel()).getId().equals(chan.getId())) || invites.isEmpty()) {
+                                if (guild.getVoiceStates().stream().anyMatch(x -> Objects.equals(x.getChannel(), chan)))
+                                    chan.createInvite().setMaxAge(3600).setMaxUses(1).setTemporary(false).setUnique(true).complete();
+                                else
+                                    chan.delete().complete();
+                            }
                         }
                     }
                 }
